Add functions to generate dummy context
There are a couple of (hacky) functions to take a template,
extract out all django {{ FOO }} variables and then generate
a dictionary containing { 'foo': "{{ FOO}"" } out of it, so that
template renders have placeholders in. It works with dotted
notation (follows down the tree), but obviously doesn't do
anything with tags or filters.